cloudform/requirements.txt
Twisted==20.3.0
aioredis==1.3.1
amqp==2.6.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
cairocffi==1.7.1
CairoSVG==2.7.1
celery==4.4.7
certifi==2024.7.4
cffi==1.16.0
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
click==7.1.2
constantly==23.10.4
cryptography==42.0.8
cssselect2==0.7.0
cx-Oracle==8.0.1
daphne==2.5.0
defusedxml==0.7.1
Django==2.2.3
django-auth-ldap==2.0.0
django-celery-beat==2.0.0
django-ckeditor==5.9.0
django-cors-headers==3.0.2
django-elasticsearch-dsl==7.1.4
django-extensions==2.2.1
django-fernet-fields==0.6
django-filter==2.2.0
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-timezone-field==4.2.3
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
docopt==0.6.2
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
flower==0.9.5
graypy==2.1.0
h11==0.8.1
hiredis==2.3.2
html5lib==1.1
httptools==0.0.13
humanize==4.9.0
hyperlink==21.0.0
idna==2.8
incremental==22.10.0
jmespath==0.9.4
kombu==4.6.11
ldap3==2.6.1
msgpack==0.6.2
Pillow==9.5.0
prometheus-client==0.8.0
prompt_toolkit==3.0.47
psycopg2-binary==2.8.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyHamcrest==2.1.0
PyJWT==1.7.1
pyOpenSSL==24.1.0
pyphen==0.15.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-ldap==3.4.4
pytz==2024.1
pyvim==3.0.2
pyvmomi==8.0.3.0.1
PyYAML==5.4.1
redis==3.5.3
requests==2.22.0
service-identity==18.1.0
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
tornado==6.4.1

txaio==23.1.1
typing_extensions==4.12.2
urllib3==1.25.11
uvicorn==0.10.8
uvloop==0.19.0
vine==1.3.0
wcwidth==0.2.13
WeasyPrint==52.2
webencodings==0.5.1
websockets==8.1
whitenoise==4.1.4
zope.interface==6.4.post2